feat(exception) : add exceptions + clean code on controller+service

+++ src/main/java/com/falkenberg/moto_repair_api/services/RepairingOrderService.java
@@ -2,12 +2,16 @@ package com.falkenberg.moto_repair_api.services;
 
 import com.falkenberg.moto_repair_api.components.UtilComponent;
 import com.falkenberg.moto_repair_api.dtos.RepairingOrderDto;
+import com.falkenberg.moto_repair_api.entities.RepairingOrder;
 import com.falkenberg.moto_repair_api.enums.Status;
+import com.falkenberg.moto_repair_api.exceptions.ApiException;
+import com.falkenberg.moto_repair_api.exceptions.ErrorCode;
 import com.falkenberg.moto_repair_api.repositories.RepairingOrderRepository;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Sort;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 
 @Service
@@ -31,11 +35,17 @@ public class RepairingOrderService {
     }
 
     public RepairingOrderDto getRepairingOrder(Long id){
+
         return utilComponent.repairingOrderEntityToDto(repairingOrderRepository.getReferenceById(id));
     }
 
     public RepairingOrderDto addRepairingOrder(RepairingOrderDto repairingOrderDto){
-        return utilComponent.repairingOrderEntityToDto(repairingOrderRepository.save(utilComponent.repairingOrderToEntity(repairingOrderDto)));
+
+        RepairingOrder repairingOrder = utilComponent.repairingOrderToEntity(repairingOrderDto);
+
+        RepairingOrder savedRepairingOrderDto = repairingOrderRepository.save(repairingOrder);
+
+        return utilComponent.repairingOrderEntityToDto(savedRepairingOrderDto);
     }
 
     public RepairingOrderDto putRepairingOrder(Long id, RepairingOrderDto repairingOrderDto){
@@ -47,6 +57,11 @@ public class RepairingOrderService {
     }
 
     public void deleteRepairingOrder(Long id) {
+
+        if(!repairingOrderRepository.existsById(id)){
+            throw new ApiException("", ErrorCode.NOT_FOUND, HttpStatus.NOT_FOUND);
+        }
+
         repairingOrderRepository.delete(repairingOrderRepository.getReferenceById(id));
     }
